refactor: Update transformQuestions and calculateSocialPreferenceAnswers functions
This commit refactors the transformQuestions function in the intro/helpers.ts file to update the calculation of the totalQuestions property. It now subtracts 1 from the length of the questions because the last item in basic questions array is the optional text.

Additionally, the calculateSocialPreferenceAnswers function in the intro/hooks/useQuestionCategories.ts file is updated to modify the singleKeys array. The keys "4", "5", "8", and "12" are now used instead of the previous keys.

These changes improve the logic and functionality of the code related to transforming questions and calculating social preference answers.